refactor(ui): improve responsive styling for cart and suggest components
- 將 textFormatting 移至 helpers/format 路徑底下
- 調整 ProductItem 元件內的 Footer、Rental 區塊的響應式佈局和樣式，優化手機和平板兩個斷點的間距、字體大小和 flex 佈局排版
