+++ helpers/format/textFormatting.ts
